[CINN Frontend] Optimize user interface, polish fuse_pass classes"
