+++ cmake/cinn/core.cmake
@@ -433,28 +433,6 @@ function(download_and_uncompress INSTALL_DIR URL FILENAME)
     INSTALL_COMMAND "")
 endfunction()
 
-set(fusion_pass_file
-    ${CMAKE_CURRENT_BINARY_DIR}/paddle/cinn/hlir/pass/use_general_pass.h
-    CACHE INTERNAL "use_general_pass.h file")
-file(
-  WRITE ${fusion_pass_file}
-  "#include \"paddle/cinn/common/macros.h\" // Generated by the paddle/cinn/hlir/pass/CMakeLists.txt.  DO NOT EDIT!\n\n"
-)
-
-function(find_fusion_pass_register FILENAME ADD_PATH PATTERN)
-  # set op_name to OUTPUT
-  file(READ ${FILENAME} CONTENT)
-  string(REGEX MATCHALL "${PATTERN}\\([a-zA-Z0-9_]*," fusion_pass_patterns
-               "${CONTENT}")
-  if(NOT fusion_pass_patterns STREQUAL "")
-    foreach(pass_pattern ${fusion_pass_patterns})
-      string(REPLACE "${PATTERN}(" "" pass_pattern "${pass_pattern}")
-      string(REPLACE "," "" pass_pattern "${pass_pattern}")
-      file(APPEND ${ADD_PATH} "USE_FUSION_PASS(${pass_pattern});\n")
-    endforeach()
-  endif()
-endfunction()
-
 function(gather_srcs SRC_GROUP)
   set(options)
   set(oneValueArgs)
@@ -464,8 +442,6 @@ function(gather_srcs SRC_GROUP)
     set(${SRC_GROUP}
         "${${SRC_GROUP}};${CMAKE_CURRENT_SOURCE_DIR}/${cpp}"
         CACHE INTERNAL "")
-    find_fusion_pass_register("${CMAKE_CURRENT_SOURCE_DIR}/${cpp}"
-                              ${fusion_pass_file} "CINN_REGISTER_FUSION_PASS")
   endforeach()
 endfunction()
